0625update

+++ server/src/main/java/org/apache/iotdb/db/wal/buffer/IWALBuffer.java
@@ -35,7 +35,7 @@ public interface IWALBuffer extends AutoCloseable {
   void write(WALEntry walEntry);
 
   /** Get current log version id */
-  int getCurrentWALFileVersion();
+  long getCurrentWALFileVersion();
 
   /** Get current wal file's size */
   long getCurrentWALFileSize();
